Long arithmetix signed fixes

+++ mod_pow.c
@@ -1,7 +1,7 @@
 #include "stdio.h"
 
 
-typedef long long ull;
+typedef long long int ull;
 
 int mod_pow_recursive(int x, long y, int m)
 {
@@ -18,7 +18,7 @@ int mod_pow_recursive(int x, long y, int m)
     }
 }
     
-int mod_pow(ull x, ull y, ull m)
+ull mod_pow(ull x, ull y, ull m)
 {
     ull ans = 1;
     while (y)
